Made infinite lives on menu
What it does, is when the player fires, it decrements the number of
sides as usual, but then if the current local scene is one of the level
select screens (if statement with or conditions) then it re-increments
the number of sides. That way there’s also a cool little animation when
you pulse out. Let me know what you guys think!